Refactors map to use flatMap
This removes calls to unsafeRun from the map. We could have implemented
flatMap in terms of map, but it would have broken the specification of
the map function that requires only FP functions (total, deterministic,
no actions) to be used for callbacks. As we're working with actions, we
would break the no actions rule by having to run with unsafeRun