Implement `zip()` on Validation
Key point is that we can pattern-match on a tuple 💡. This allows us to
deal with every Valid/Invalid permutation sequentially, rather than in
nested matches.

As `zip` returns a tuple of the return values, we discard them for the
checking the username, but are required to map across the result (i.e.
happy path) to decalre that we're creating a username. On validating for
a User, it's simpler as we pass the validated Country and Username into
the User object in the body of the function passed to `map`